hpsa: use driver instead of system work queue for command resubmission
There is a possibility of deadlock if we use the system work queue
for command resubmission since something in the queue may be depending
on the i/o that gets resubmitted, and the resubmitted i/o will be
behind the thing that depends on it in the queue.  Using a driver
specific work queue avoids this.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>