hpsa: do not wait for aborted commands to complete after abort
After an abort command completes successfully there is no need
or sense in trying to wait for the aborted command to complete.
It will have already completed (either normally, or as aborted)
before the abort command completes.  Trying to wait for it to
complete later doesn't make sense -- it would only be waiting for
the command to be re-used for something else and for *that* to
complete, which would be crazy.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>