hpsa: add masked physical devices into h->dev[] array
This is so that we can store the ioaccel handle for those
devices so that when we need to abort commands sent down
ioaccel2 path, we can look up the 8-byte LUN ID in h->dev[]
instead of having to do i/o to the controller each time.

We add an expose_state field to h->dev[] elements to indicate
how the device is exposed to the scsi mid layer: Not at all,
without an upper level driver (no_uld_attach) or
normally exposed.

Signed-off-by: Joe Handzik <joseph.t.handzik@hp.com>
Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>
Reviewed-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>