hpsa: avoid unnecessary io in hpsa_get_pdisk_of_ioaccel2()
THIS PATCH IS PROBLEMATIC - DO NOT SUBMIT UPSTREAM AS IS.

The problem is that h->dev[] does not contain data for unexposed
disks, which is what we need for getting the physical disk of
an io by matching up the ioaccel handle.  A later patch gets
this data and puts it in h->dev[] so this patch eventually works,
but not at this point in the stack.   Below is the original
commit message.

Instead of doing CISS_REPORT_PHYSICAL to get the LUNID for
the physical disk in hpsa_get_pdisk_of_ioaccel2(), just get
it out of h->dev[] where we already have it cached.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>