hpsa: limit number of concurrent abort requests
If all available commands are outstanding and the abort handler
is invoked, the abort handler may not be able to allocate a command
and may busy-wait excessivly.  Reserve a small number of commands
for the abort handler and limit the number of concurrent abort
requests to the number of reserved commands.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>