hpsa:  second round of tagged-allocation review-input changes
Remove the reserved_cmds commands field from the controller info
structure and replace references to it with a #define'd symbol.  Add a
build-time assertion that the minimum number of available commands is
greater than the number which are reserved by the driver to ensure
that we can do at least some I/O.  Replace the magic number used for
the minimum number of commands with at #define'd symbol.

(This patch should presumably be squashed with the original patch and
the previous review-input changes.)

Signed-off-by: Webb Scales <webbnh@hp.com>