hpsa: get rid of cmd_special_alloc and cmd_special_free
Always use cmd_alloc and cmd_free instead, we should have
reserved enough commands that we should not need to ever
dynamically allocate any and should always have enough.

This is laying the groundwork for removing the internal
queue of commands from the driver so that the locks that
protect that queue may be removed.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>