hpsa: factor out hpsa_ciss_submit function
Factor out the bottom part of the queuecommand function
which is the part that builds commands for submitting down
the "normal' RAID stack path of a Smart Array.

Need to factor this out to improve how commands that
were initially sent down one of the "ioaccellerated"
paths but which have some sort of error condition are
retried down the "normal" path.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>