hpsa: fix race when updating raid map data.
When enabling io accelerator paths for a logical drive make sure to
update the raid map data prior to setting the ioaccel_enabled flag
for that logical drive, and when disabling ioaccelerator path for a
logical drive, do not update the raid map data at all.  One of the
reasons that the ioaccel_enabled flag might not be set is because
getting the raid map data failed, in which case, updating the raid
map data is certainly the wrong thing to do.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>