hpsa: allow commands to be completed on specified reply queue
This is to enable fixing a race between command completions and
abort completions on different reply queues in a subsequent patch.
We want to be able to specify which reply queue an abort completion
should occur on so that it cannot race the completion of the command
it is trying to abort.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>