hpsa: do not send two aborts with swizzled tags.
Some Smart Arrays required aborts to be sent with tags in
the wrong endian byte order.  To avoid having to know about
this, we would send two aborts with tags with each endian order.
On high IOPS devices, this turns out to be not such a hot idea.
So we now have a list of the devices that got the tag backwards,
and we only send it one way.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>