hpsa: add masked physical devices into h->dev[] array
This is so that we can store the ioaccel handle for those
devices so that when we need to abort commands sent down
ioaccel2 path, we can look up the 8-byte LUN ID in h->dev[]
instead of having to do i/o to the controller each time.

We add an expose_state field to h->dev[] elements to indicate
how the device is exposed to the scsi mid layer: Not at all,
without an upper level driver (no_uld_attach) or
normally exposed.

Since masked physical devices are now present in h->dev[] array
it is perfectly possible to do

	echo scsi add-single-device 2 2 0 0 > /proc/scsi/scsi

and bring them online.  This should not be the case for masked
physical devices..

Change the values for the expose_status field of hpsa_scsi_dev_t to make sense.

HPSA_DO_NOT_EXPOSE     0x0
HPSA_SG_ATTACH         0x1
HPSA_ULD_ATTACH        0x2
HPSA_SCSI_ADD          (HPSA_SG_ATTACH | HPSA_ULD_ATTACH)

Signed-off-by: Joe Handzik <joseph.t.handzik@hp.com>
Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>
Reviewed-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>