hpsa: fix code that updates h->dev[]->phys_disk[]
h->dev[]->phys_disk[] contains a map of the physical
disks that make up each logical drive.  When a RAID
migration occurs, this map can change for a logical
drive.  We need to make sure it gets updated in a
consistent way and that this mapping is not used until
after it is finished being updated.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>