hpsa:  use block layer tag for command allocation
Modify cmd_tagged_allocation() to crash when it receives a bad tag from the
block layer, instead of returning a NULL pointer and letting the next layer
up fail:  we should never get a bad tag, and if we do it means that the
system is misconfigured (e.g., that the scsi-mq support is not enabled),
and it would be best to die a clean, swift death before anyone gets hurt.

Limit the search for free cmd blocks to only the reserved range, and remove the
optimization which avoids starting the search at the beginning of the range.

Remove the reserved_cmds commands field from the controller info
structure and replace references to it with a #define'd symbol.  Add a
build-time assertion that the minimum number of available commands is
greater than the number which are reserved by the driver to ensure
that we can do at least some I/O.  Replace the magic number used for
the minimum number of commands with at #define'd symbol.

reserve part of the tag space for the driver

move tag selection to a helper function

If the blk-mq support isn't available, set up the SCSI tagged command
queuing (tcq) and use the SCSI midlayer tag.

Signed-off-by: Webb Scales <webbnh@hp.com>
Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>