hpsa: do not poll device with TUR after device reset
If a logical drive has failed, resetting it will ensure
outstanding commands are completed, but polling it with
TURs after the reset will not work because the TURs will
never report good status.  So successful TUR should not
be a condition of success for the device reset error
handler.

NOTE: THERE IS A PROBLEM WITH THIS PATCH.  While it fixes
the problem that a failed logical drive will never become
ready, and so fixes the problem that the reset device error
handler would get stuck for too long waiting for a failed
logical drive to become ready, it defeats the purpose of
that barrage of test unit readys, which was to avoid races
of the completions of aborted commands with the device reset
command completion in the completion queues.  So, an alternative
solution for those races is needed (webb is working on this).

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>